+++ docs/CHANGE_LOG.md
@@ -11,6 +11,7 @@
 - Added new public function "getAlias()", which is used to get a language alias.
 - Added new public function "getAliases()", which is used to get all language aliases.
 - Renamed the public function "getAllLanguages" to "getLanguages()".
+- Renamed the public function "getAllElementsHighlighted()" to "getElementsHighlighted()".
 
 #### **Documentation:**
 - More documentation improvements to show how some of the new features are used.